requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.4
django-ninja==1.4.3
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
